requirements.txt
flask==3.1.0
line-bot-sdk>=3.0.0
requests==2.32.3
beautifulsoup4==4.12.3
apscheduler==3.10.4
pytz==2024.2
gunicorn==23.0.0
pandas==2.2.3
openpyxl==3.1.5
psycopg2-binary==2.9.10
